requirements.txt
altgraph==0.17.4
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.7
iniconfig==2.0.0
macholib==1.16.3
modulegraph==0.19.6
packaging==24.1
pefile==2023.2.7
pluggy==1.5.0
py2app==0.28.8
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
PyQt6==6.7.0
PyQt6-Qt6==6.7.1
PyQt6-sip==13.6.0
pyqt6-tools==6.4.2.3.3
PySide6_Essentials==6.7.1
pytz==2024.1
pywin32-ctypes==0.2.2
requests==2.32.3
setuptools==70.0.0
shiboken6==6.7.1
urllib3==2.2.1
